API/requirements.txt
requests==2.22.0
google-api-core==2.3.2
google-auth==2.3.3
google-cloud-language==2.3.1
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
pymongo==4.0.1
